[REFACTOR] Manager > Entities 로 폴더 명 변경
