+++ SaveEarthTuist/Projects/SaveEarth/SaveEarth/Sources/Entities/Weather/Model/TemperatureData.swift
